BugFixed

+++ Aban360.ReportPool.Domain/Features/BuiltIns/ServiceLinkTransaction/Outputs/FinancialStatementDataOutputDto.cs
@@ -3,17 +3,17 @@
     public record FinancialStatementDataOutputDto
     {
         public string UsageGroupTitle { get; set; }
-        public int CustomerCount { get; set; }
-        public int ConsumptionTotalUnit { get; set; }
-        public int DailyAverage { get; set; }
-        public int NetConsumption { get; set; }
+        public long CustomerCount { get; set; }
+        public long ConsumptionTotalUnit { get; set; }
+        public decimal DailyAverage { get; set; }
+        public long NetConsumption { get; set; }
         public long NetAmount { get; set; }
-        public int ReturnedConsumption { get; set; }
+        public long ReturnedConsumption { get; set; }
         public long ReturnedAmount { get; set; }
         public long DiscountAmount { get; set; }
-        public int RawConsumption { get; set; }
-        public long RawAmount { get; set; }
-        public float RawAmountAverage { get; set; }
-        public float ConsumptionAverageInMonth { get; set; }
+        public long RawConsumption { get; set; }
+        public decimal RawAmount { get; set; }
+        public decimal RawAmountAverage { get; set; }
+        public decimal ConsumptionAverageInMonth { get; set; }
     }
 }
